requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pandas==2.2.3
requests==2.32.3
openpyxl==3.1.5
pyyaml==6.0.2
anthropic==0.42.0
python-dotenv==1.0.1
httpx==0.28.1
